Consolidate email tests
